Move EventLoopTask into a separate file because it causes clangd to crash
